+++ demo/src/main/java/com/example/demo/dto/response/MonthlyStatsDto.java
@@ -1,9 +1,10 @@
 package com.example.demo.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record MonthlyStatsDto(
     LocalDateTime month,
-    double revenueOfTheMonth,
+    BigDecimal revenueOfTheMonth,
     Long bookingsOfTheMonth
 ) {}
